Updated the profile to show related runs and groups
