+++ routes/data/login.json
@@ -1,0 +1,1 @@
+{"id":1,"first_name":"Allison","last_name":"Melendez","email":"allisonmelendez@cinaster.com","password":"doquisesse","age":20,"desired_pace":"14:16","city":"Choctaw","state":"Utah","zipcode":14893,"phone":"(837)463-2366","admin":1,"races":[{"pace":"12:30","ranking":558,"name":"Avenetro","date":"Wed May 14 2003 14:00:57 GMT-0500 (Central Daylight Time)","distance":6.0065,"email":"hermanjones@avenetro.com","phone":"(825)529-3057","city":"Waukeenah","state":"Hawaii","zipcode":91366,"race_id":6}],"running_groups":[{"rg_id":11,"name":"Freakin","phone":"(885) 407-2660","pace":"11:53","city":"Fillmore","state":"Wisconsin","zipcode":95005}]}
