+++ routes/race.js
@@ -1,35 +1,13 @@
 const express = require("express");
 const router = express.Router();
 const {Race} = require("../db/ready_race_run");
-
-function getStates(races){
-  states = []
-
-  for (let i = 0; i < races.length; i++){
-    states.push({id: races[i].id, s: races[i].state})
-  }
-  return states
-}
-
-function getFiltered(results, query, state){
-  if (state == "All" && query.trim() == ""){
-    return results
-  } 
-  output = []
-  for (let i = 0; i < results.length; i++){
-    console.log(typeof results[i].name)
-    if ((state != "All" && results[i].state == state)||results[i].name.includes(query)){
-      output.push(results[i])
-    }
-  }
-  return output
-}
+const {getStates, getFiltered} = require("../views/public/javascript/searchResults")
 
 router.get("/", (req, res) => {
      Race.fetchAll()
     .then( Race => {
       result = JSON.parse(JSON.stringify( Race))
-      res.status(200).render('viewraces', { data: getFiltered(result, "", "All"), states: getStates(result) });
+      res.status(200).render('viewraces', { query: "", state: "", data: getFiltered(result, "", "All"), states: getStates(result) });
     })
     .catch(err => {
         console.log(err);
@@ -41,7 +19,7 @@ router.post("/", (req, res) => {
   Race.fetchAll()
   .then( races => {
     result = JSON.parse(JSON.stringify( races))
-    res.status(200).render('viewraces', { data: getFiltered(result, req.body.query, req.body.selectpicker), states: getStates(result) });
+    res.status(200).render('viewraces', { query: req.body.query, state: req.body.selectpicker, data: getFiltered(result, req.body.query, req.body.selectpicker), states: getStates(result) });
   })
   .catch(err => {
     console.log(err);
