+++ routes/running_group.js
@@ -1,35 +1,17 @@
 const express = require("express");
 const router = express.Router();
 const {RunningGroup} = require("../db/ready_race_run");
+const {getStates, getFiltered} = require("../views/public/javascript/searchResults")
 
-function getStates(runningGroups){
-  states = []
-
-  for (let i = 0; i < runningGroups.length; i++){
-    states.push({id: runningGroups[i].id, s: runningGroups[i].state})
-  }
-  return states
-}
-
-function getFiltered(results, query, state){
-  if (state == "All" && query.trim() == ""){
-    return results
-  } 
-  output = []
-  for (let i = 0; i < results.length; i++){
-    console.log(typeof results[i].name)
-    if ((state != "All" && results[i].state == state)||results[i].name.includes(query)){
-      output.push(results[i])
-    }
-  }
-  return output
-}
 
 router.get("/", (req, res) => {
   RunningGroup.fetchAll()
  .then( RunningGroup => {
    result = JSON.parse(JSON.stringify( RunningGroup))
-   res.status(200).render('allrunninggroups', { data: getFiltered(result, "", "All"), states: getStates(result) });
+   res.status(200).render('allrunninggroups', { query: "", 
+        state: "", 
+        data: getFiltered(result, "", "All"), 
+        states: getStates(result) });
  })
  .catch(err => {
      console.log(err);
@@ -41,7 +23,10 @@ router.post("/", (req, res) => {
   RunningGroup.fetchAll()
   .then( runningGroups => {
     result = JSON.parse(JSON.stringify( runningGroups))
-    res.status(200).render('allrunninggroups', { data: getFiltered(result, req.body.query, req.body.selectpicker), states: getStates(result) });
+    res.status(200).render('allrunninggroups', { query: req.body.query, 
+      state: req.body.selectpicker, 
+      data: getFiltered(result, req.body.query, req.body.selectpicker), 
+      states: getStates(result) });
   })
   .catch(err => {
     console.log(err);
