+++ routes/user.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const router = express.Router();
 const {User, RaceHistory, Race, RunningGroup, RunningGroupMember} = require("../db/ready_race_run");
+const {Login} = require("./data/userData");
+const fs = require('fs');
 
 create = (req, res) => {
     User.forge( req.body )
@@ -33,8 +35,11 @@ findByEmail = (req, res) => {
         if (!User) {
             res.status(200).render('index', { });
         } else {
-            console.log(User)
-            res.status(200).render('profile', { data: JSON.parse(JSON.stringify(User)), loggedIn: true });
+            if (!User) {
+                res.render('index', {data:{}})
+            } else {
+                res.status(200).render('profile', { data: Login(User), races: User.races, rgs: User.running_groups});
+            }
         }
     })
     .catch(err => {
@@ -44,73 +49,13 @@ findByEmail = (req, res) => {
 };
 
 findById = (req, res) => {
-    console.log("Made it here")
     User.forge().query({where:{ id: req.params.id}})
     .fetch()
-    .then((User) => {
+    .then(async function(User)  {
         if (!User) {
-            res.render('index', {})
+            res.render('index', {data:{}})
         } else {
-            hist_res = []
-            rgs = []
-            RunningGroupMember.forge().query({where: {user_id: User.id}}).fetchAll().then(
-                groups => {
-                    const result = []
-                    for(let i = 0; i < groups.models.length; i++){
-                        RunningGroup.forge().query({where: {id: groups.models[i].attributes.running_group_id}}).fetch().then(
-                            rg => {
-                                let r = {
-                                    rg_id: rg.attributes.id,
-                                    name: rg.attributes.name,
-                                    phone: rg.attributes.phone,
-                                    pace: rg.attributes.pace,
-                                    city: rg.attributes.city,
-                                    state: rg.attributes.state,
-                                    zipcode: rg.attributes.zipcode
-                                }
-                                hist_res.push(r)
-                            }
-                        )
-                    }
-                }
-            )
-            RaceHistory.forge().query({where: {user_id: User.id}}).fetchAll().then(
-                history => {
-                    for(let i = 0; i < history.models.length; i++){
-                        Race.forge().query({where: {id: history.models[i].attributes.race_id}}).fetch().then(
-                            race => {
-                                let r = {
-                                    pace: history.models[i].attributes.pace, 
-                                    ranking: history.models[i].attributes.ranking,
-                                    name: race.attributes.name,
-                                    date: race.attributes.date,
-                                    distance: race.attributes.distance,
-                                    email: race.attributes.email,
-                                    phone: race.attributes.phone,
-                                    city: race.attributes.city,
-                                    state: race.attributes.state,
-                                    zipcode: race.attributes.zipcode,
-                                    race_id: race.attributes.id
-                                }
-                                hist_res.push(r)
-                            }
-                        ).catch(err => {
-                            console.log(err);
-                            res.status(500).json(err);
-                        })
-                    }
-                    
-                    
-                    res.status(200).render('profile', { data: JSON.parse(JSON.stringify(User)), races: hist_res, rgs: rgs});
-                    // races = JSON.parse(JSON.stringify(RaceHistory))
-                    // Race.forge().query({where: {id: RaceHistory.race_id}}).fetch().then(race => {
-                    //     res.status(200).render('profile', { data: JSON.parse(JSON.stringify(User)), races: race });
-                    // })
-                }
-            ).catch(err => {
-                console.log(err);
-                res.status(500).json(err);
-            })
+            Login(res, User)
         }
     })
     .catch(err => {
