+++ views/index.hbs
@@ -4,7 +4,7 @@
 
     <div class="col-lg-6 text-block">
         <h3>What We Do </h3>
-        <p>Running races are growing in popularity every year. Whether your goal is to run a 5k, 10k, half marathon or marathon, there are several choices out there. With all these choices, finding the right race for you becomes difficult. Our goal is to construct a website that will help a user find a future race and view all previous races.</p>
+        <p>Running races are growing in popularity every year. Whether your goal is to run a 5k, 10k, half marathon or marathon, there are several choices out there. With all these choices, finding the right race for you becomes difficult. Our goal is to give you an experience to find a future race and view all previous races.</p>
         
     </div>
 </section>
@@ -20,9 +20,8 @@
             </div>
             <div class="col-lg-6 order-lg-1">
                 <div class="p-5">
-                    <h2 class="display-4">Find race groups in your area ..</h2>
-                    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Quod aliquid, mollitia odio veniam sitiste esse assumenda amet aperiam exercitationem, ea animi blanditiis recusandae! Ratione
-                        voluptatum molestiae adipisci, beatae obcaecati.</p>
+                    <h2 class="display-4">Find race groups!</h2>
+                    <p>Find people who run at your pace at a place near you! Traveling? No problem! Search by zipcode! Want people at your pace? Search by pace!</p>
                 </div>
             </div>
         </div>
@@ -39,35 +38,10 @@
             </div>
             <div class="col-lg-6">
                 <div class="p-5">
-                    <h2 class="display-4">Connect with people like you!</h2>
-                    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Quod aliquid, mollitia odio veniam sit
-                        iste esse assumenda amet aperiam exercitationem, ea animi blanditiis recusandae! Ratione
-                        voluptatum molestiae adipisci, beatae obcaecati.</p>
+                    <h2 class="display-4">Find races!</h2>
+                    <p>Find races at a time, place and distance that works with you! See all of your previous races!</p>
                 </div>
             </div>
         </div>
     </div>
 </section>
-
-<section>
-    <div class="container">
-        <div class="row align-items-center">
-            <div class="col-lg-6 order-lg-2">
-                <div class="p-5">
-                    <img class="img-fluid rounded-circle" src="https://media.gettyimages.com/photos/time-to-clock-another-personal-best-picture-id916060452" alt="">
-                </div>
-            </div>
-            <div class="col-lg-6 order-lg-1">
-                <div class="p-5">
-                    <h2 class="display-4">Get recommendations!</h2>
-                    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Quod aliquid, mollitia odio veniam sit
-                        iste esse assumenda amet aperiam exercitationem, ea animi blanditiis recusandae! Ratione
-                        voluptatum molestiae adipisci, beatae obcaecati.</p>
-                </div>
-            </div>
-        </div>
-    </div>
-
-</section>  
-
-
