+++ views/main.hbs
@@ -32,7 +32,11 @@
                         <a class="nav-link" href="/running_group">Running Group</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="/user/create" data-toggle="modal" data-target=".log-sign">Login</a>
+                        {{#if data}}
+                            <a class="nav-link" href="/user/profile/{{data.id}}">{{data.first_name}} {{data.last_name}}</a>
+                        {{else}}
+                            <a class="nav-link" href="/user/create" data-toggle="modal" data-target=".log-sign">Login</a>
+                        {{/if}}
                     </li>
                 </ul>
             </div>
@@ -151,16 +155,18 @@
     <div class="container">
         <div class="row">
             <div class="col-sm-4 myCols">
-                <h5>Get started</h5>
+
                 <ul>
+                    <li><h5>Get started</h5></li>
                     <li><a href="/user/create">Home</a></li>
                     <li><a href="/race">Races</a></li>
                     <li><a href="/running_group">Running Groups</a></li>
                 </ul>
             </div>
             <div class="col-sm-4 myCols">
-                <h5>About us</h5>
+                
                 <ul>
+                    <li><h5>About us</h5></li>
                     <li><a target="_blank" href="https://github.com/elizabeth-phillips">Elizabeth
                             Phillips</a></li>
                     <li><a target="_blank" href="https://github.com/shwetarane">Shweta Rane</a></li>
@@ -168,8 +174,9 @@
                 </ul>
             </div>
             <div class="col-sm-4 myCols">
-                <h5>Contact Us</h5>
+                
                 <ul>
+                    <li><h5>Contact Us</h5></li>
                     <li>Texas State University</li>
                     <li>601 University Dr</li>
                     <li>San Marcos, TX 78666</li>
